// buttazzoEsame/src/types/alunno.ts
export interface Alunno {
  cognome: string
  nome: string
  comuneNascita: string
  provNascita: string
  dataNascita: string
  capResidenza: string
  comuneResidenza: string
  provResidenza: string
  indirizzoResidenza: string
  emailPrincipale: string
  cell: string
  codiceFiscale: string
  matricola: string
  emailSecondaria: string
  foto: string
}